completed expression

+++ circle.py
@@ -1,3 +1,3 @@
 radius_input = input("Enter the radius: ")
-area = 3.14 * float(radius_input)**2
+area = 3.14 * int(radius_input)**2
 print(("Area of a circle: " + str(area) + " square units"))
